refac,docs: args,config: Make "max pixels" opt-in
- Change: Include zero in the value range for `max pixels` config
  option and `--max-pixels` CL option.

  A value of zero implies the absence of a maximum pixel-count i.e
  all images are rendered regardless of their resolution.

- Change: Change the default value of `max pixels` config option to
  zero.
- Change: Update the documentation of the `max pixels` config option
  and `--max-pixels` CL option.

  - Update the description, value range and default value.
  - Remove the admonition about performance implications.

- Change: Remove the footnote for the `--max-pixels` CL option; the
  repetition is unnecessary.

Turns out displaying all images regardless of resolution results in a
better out-of-the-box user experience as that's the expectation of
literally every user (other than myself :facepalm:).
Also, the so-called "performance implications" of displaying high-res
images turns to be really not as bad as I had it pictured in my mind.

Thanks to [@qrockz](https://github.com/qrockz) and
[@DreamMaoMao](https://github.com/DreamMaoMao) for helping me make the
right decision :smiley:.

Refs: #12.